Tests: separate backend and core fun + speed test for SVD
